rest of pieces (all but setActiveEventThunk)
