+++ apps/studio/src/store/slices/CFS/error.ts
@@ -1,10 +1,10 @@
-/* eslint-disable import/prefer-default-export */
-/* eslint-disable max-classes-per-file */
-class CodemodRuntimeFailureError extends Error {
-	constructor() {
-		// @TODO better description
-		super("Codemod has runtime errors");
-	}
-}
-
-export { CodemodRuntimeFailureError };
+// /* eslint-disable import/prefer-default-export */
+// /* eslint-disable max-classes-per-file */
+// class CodemodRuntimeFailureError extends Error {
+// 	constructor() {
+// 		// @TODO better description
+// 		super("Codemod has runtime errors");
+// 	}
+// }
+//
+// export { CodemodRuntimeFailureError };
